+++ src/libAtomVM/opcodesswitch.h
@@ -1920,7 +1920,6 @@ schedule_in:
                 break;
             }
 
-            //TODO: implement me
             case OP_BIF1: {
                 uint32_t fail_label;
                 DECODE_LABEL(fail_label, pc);
@@ -1945,7 +1944,12 @@ schedule_in:
                     DEBUG_FAIL_NULL(func);
                     term ret = func(ctx, arg1);
                     if (UNLIKELY(term_is_invalid_term(ret))) {
-                        HANDLE_ERROR();
+                        if (fail_label) {
+                            pc = mod->labels[fail_label];
+                            break;
+                        } else {
+                            HANDLE_ERROR();
+                        }
                     }
 
                     WRITE_REGISTER(dreg, ret);
@@ -1953,7 +1957,6 @@ schedule_in:
                 break;
             }
 
-            //TODO: implement me
             case OP_BIF2: {
                 uint32_t fail_label;
                 DECODE_LABEL(fail_label, pc);
@@ -1981,7 +1984,12 @@ schedule_in:
                     DEBUG_FAIL_NULL(func);
                     term ret = func(ctx, arg1, arg2);
                     if (UNLIKELY(term_is_invalid_term(ret))) {
-                        HANDLE_ERROR();
+                        if (fail_label) {
+                            pc = mod->labels[fail_label];
+                            break;
+                        } else {
+                            HANDLE_ERROR();
+                        }
                     }
 
                     WRITE_REGISTER(dreg, ret);
@@ -5026,8 +5034,8 @@ wait_timeout_trap_handler:
             }
 
             case OP_GC_BIF1: {
-                uint32_t f_label;
-                DECODE_LABEL(f_label, pc);
+                uint32_t fail_label;
+                DECODE_LABEL(fail_label, pc);
                 uint32_t live;
                 DECODE_LITERAL(live, pc);
                 uint32_t bif;
@@ -5040,7 +5048,12 @@ wait_timeout_trap_handler:
                     GCBifImpl1 func = EXPORTED_FUNCTION_TO_GCBIF(exported_bif)->gcbif1_ptr;
                     term ret = func(ctx, live, arg1);
                     if (UNLIKELY(term_is_invalid_term(ret))) {
-                        HANDLE_ERROR();
+                        if (fail_label) {
+                            pc = mod->labels[fail_label];
+                            break;
+                        } else {
+                            HANDLE_ERROR();
+                        }
                     }
                 #endif
 
@@ -5048,27 +5061,25 @@ wait_timeout_trap_handler:
                 DECODE_DEST_REGISTER(dreg, pc);
 
                 #ifdef IMPL_EXECUTE_LOOP
-                    TRACE("gc_bif1/5 fail_lbl=%i, live=%i, bif=%i, arg1=0x%lx, dest=%c%i\n", f_label, live, bif, arg1, T_DEST_REG(dreg));
+                    TRACE("gc_bif1/5 fail_lbl=%i, live=%i, bif=%i, arg1=0x%lx, dest=%c%i\n", fail_label, live, bif, arg1, T_DEST_REG(dreg));
                     WRITE_REGISTER(dreg, ret);
                 #endif
 
                 #ifdef IMPL_CODE_LOADER
                     TRACE("gc_bif1/5\n");
 
-                    UNUSED(f_label)
+                    UNUSED(fail_label)
                     UNUSED(live)
                     UNUSED(bif)
                     UNUSED(arg1)
                     UNUSED(dreg)
                 #endif
-
-                UNUSED(f_label)
                 break;
             }
 
             case OP_GC_BIF2: {
-                uint32_t f_label;
-                DECODE_LABEL(f_label, pc);
+                uint32_t fail_label;
+                DECODE_LABEL(fail_label, pc);
                 uint32_t live;
                 DECODE_LITERAL(live, pc);
                 uint32_t bif;
@@ -5083,7 +5094,12 @@ wait_timeout_trap_handler:
                     GCBifImpl2 func = EXPORTED_FUNCTION_TO_GCBIF(exported_bif)->gcbif2_ptr;
                     term ret = func(ctx, live, arg1, arg2);
                     if (UNLIKELY(term_is_invalid_term(ret))) {
-                        HANDLE_ERROR();
+                        if (fail_label) {
+                            pc = mod->labels[fail_label];
+                            break;
+                        } else {
+                            HANDLE_ERROR();
+                        }
                     }
                 #endif
 
@@ -5091,22 +5107,20 @@ wait_timeout_trap_handler:
                 DECODE_DEST_REGISTER(dreg, pc);
 
                 #ifdef IMPL_EXECUTE_LOOP
-                    TRACE("gc_bif2/6 fail_lbl=%i, live=%i, bif=%i, arg1=0x%lx, arg2=0x%lx, dest=%c%i\n", f_label, live, bif, arg1, arg2, T_DEST_REG(dreg));
+                    TRACE("gc_bif2/6 fail_lbl=%i, live=%i, bif=%i, arg1=0x%lx, arg2=0x%lx, dest=%c%i\n", fail_label, live, bif, arg1, arg2, T_DEST_REG(dreg));
                     WRITE_REGISTER(dreg, ret);
                 #endif
 
                 #ifdef IMPL_CODE_LOADER
                     TRACE("gc_bif2/6\n");
 
-                    UNUSED(f_label)
+                    UNUSED(fail_label)
                     UNUSED(live)
                     UNUSED(bif)
                     UNUSED(arg1)
                     UNUSED(arg2)
                     UNUSED(dreg)
                 #endif
-
-                UNUSED(f_label)
                 break;
             }
 
@@ -5131,9 +5145,11 @@ wait_timeout_trap_handler:
                 break;
             }
 
+            // TODO: This opcode is currently uncovered by tests.
+            // We need to implement GC bifs with arity 3, e.g. binary_part/3.
             case OP_GC_BIF3: {
-                uint32_t f_label;
-                DECODE_LABEL(f_label, pc);
+                uint32_t fail_label;
+                DECODE_LABEL(fail_label, pc);
                 uint32_t live;
                 DECODE_LITERAL(live, pc);
                 uint32_t bif;
@@ -5150,7 +5166,12 @@ wait_timeout_trap_handler:
                     GCBifImpl3 func = EXPORTED_FUNCTION_TO_GCBIF(exported_bif)->gcbif3_ptr;
                     term ret = func(ctx, live, arg1, arg2, arg3);
                     if (UNLIKELY(term_is_invalid_term(ret))) {
-                        HANDLE_ERROR();
+                        if (fail_label) {
+                            pc = mod->labels[fail_label];
+                            break;
+                        } else {
+                            HANDLE_ERROR();
+                        }
                     }
                 #endif
 
@@ -5158,14 +5179,14 @@ wait_timeout_trap_handler:
                 DECODE_DEST_REGISTER(dreg, pc);
 
                 #ifdef IMPL_EXECUTE_LOOP
-                    TRACE("gc_bif3/7 fail_lbl=%i, live=%i, bif=%i, arg1=0x%lx, arg2=0x%lx, arg3=0x%lx, dest=%c%i\n", f_label, live, bif, arg1, arg2, arg3, T_DEST_REG(dreg));
+                    TRACE("gc_bif3/7 fail_lbl=%i, live=%i, bif=%i, arg1=0x%lx, arg2=0x%lx, arg3=0x%lx, dest=%c%i\n", fail_label, live, bif, arg1, arg2, arg3, T_DEST_REG(dreg));
                     WRITE_REGISTER(dreg, ret);
                 #endif
 
                 #ifdef IMPL_CODE_LOADER
                     TRACE("gc_bif2/6\n");
 
-                    UNUSED(f_label)
+                    UNUSED(fail_label)
                     UNUSED(live)
                     UNUSED(bif)
                     UNUSED(arg1)
@@ -5173,8 +5194,6 @@ wait_timeout_trap_handler:
                     UNUSED(arg3)
                     UNUSED(dreg)
                 #endif
-
-                UNUSED(f_label)
                 break;
             }
 
@@ -5648,11 +5667,21 @@ wait_timeout_trap_handler:
                     ctx->fr[freg3] = ctx->fr[freg1] + ctx->fr[freg2];
                     #ifdef HAVE_PRAGMA_STDC_FENV_ACCESS
                         if (fetestexcept(FE_OVERFLOW)) {
-                            RAISE_ERROR(BADARITH_ATOM);
+                            if (fail_label) {
+                                // Not sure this can happen, float operations
+                                // in guards are translated to gc_bif calls
+                                pc = mod->labels[fail_label];
+                            } else {
+                                RAISE_ERROR(BADARITH_ATOM);
+                            }
                         }
                     #else
                         if (!isfinite(ctx->fr[freg3])) {
-                            RAISE_ERROR(BADARITH_ATOM);
+                            if (fail_label) {
+                                pc = mod->labels[fail_label];
+                            } else {
+                                RAISE_ERROR(BADARITH_ATOM);
+                            }
                         }
                     #endif
                 #endif
@@ -5685,11 +5714,21 @@ wait_timeout_trap_handler:
                     ctx->fr[freg3] = ctx->fr[freg1] - ctx->fr[freg2];
                     #ifdef HAVE_PRAGMA_STDC_FENV_ACCESS
                         if (fetestexcept(FE_OVERFLOW)) {
-                            RAISE_ERROR(BADARITH_ATOM);
+                            if (fail_label) {
+                                // Not sure this can happen, float operations
+                                // in guards are translated to gc_bif calls
+                                pc = mod->labels[fail_label];
+                            } else {
+                                RAISE_ERROR(BADARITH_ATOM);
+                            }
                         }
                     #else
                         if (!isfinite(ctx->fr[freg3])) {
-                            RAISE_ERROR(BADARITH_ATOM);
+                            if (fail_label) {
+                                pc = mod->labels[fail_label];
+                            } else {
+                                RAISE_ERROR(BADARITH_ATOM);
+                            }
                         }
                     #endif
                 #endif
@@ -5722,11 +5761,21 @@ wait_timeout_trap_handler:
                     ctx->fr[freg3] = ctx->fr[freg1] * ctx->fr[freg2];
                     #ifdef HAVE_PRAGMA_STDC_FENV_ACCESS
                         if (fetestexcept(FE_OVERFLOW)) {
-                            RAISE_ERROR(BADARITH_ATOM);
+                            if (fail_label) {
+                                // Not sure this can happen, float operations
+                                // in guards are translated to gc_bif calls
+                                pc = mod->labels[fail_label];
+                            } else {
+                                RAISE_ERROR(BADARITH_ATOM);
+                            }
                         }
                     #else
                         if (!isfinite(ctx->fr[freg3])) {
-                            RAISE_ERROR(BADARITH_ATOM);
+                            if (fail_label) {
+                                pc = mod->labels[fail_label];
+                            } else {
+                                RAISE_ERROR(BADARITH_ATOM);
+                            }
                         }
                     #endif
                 #endif
@@ -5759,11 +5808,21 @@ wait_timeout_trap_handler:
                     ctx->fr[freg3] = ctx->fr[freg1] / ctx->fr[freg2];
                     #ifdef HAVE_PRAGMA_STDC_FENV_ACCESS
                         if (fetestexcept(FE_OVERFLOW | FE_DIVBYZERO)) {
-                            RAISE_ERROR(BADARITH_ATOM);
+                            if (fail_label) {
+                                // Not sure this can happen, float operations
+                                // in guards are translated to gc_bif calls
+                                pc = mod->labels[fail_label];
+                            } else {
+                                RAISE_ERROR(BADARITH_ATOM);
+                            }
                         }
                     #else
                         if (!isfinite(ctx->fr[freg3])) {
-                            RAISE_ERROR(BADARITH_ATOM);
+                            if (fail_label) {
+                                pc = mod->labels[fail_label];
+                            } else {
+                                RAISE_ERROR(BADARITH_ATOM);
+                            }
                         }
                     #endif
                 #endif
@@ -5939,7 +5998,11 @@ wait_timeout_trap_handler:
 
                     if (!(term_is_binary(src) || term_is_match_state(src))) {
                         WRITE_REGISTER(dreg, src);
-                        pc = mod->labels[fail];
+                        if (term_is_integer(fail)) {
+                            pc = mod->labels[term_to_int(fail)];
+                        } else {
+                            RAISE_ERROR(BADARG_ATOM);
+                        }
                     } else {
                         term match_state = term_alloc_bin_match_state(src, 0, &ctx->heap);